# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Homium Compliance Portal API — templates, versions, saved sections, applications
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /applications:
    get:
      operationId: listApplications
      tags: [intake]
      summary: All applications with progress summary
      responses:
        "200":
          description: Applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApplicationSummary"
    post:
      operationId: createApplication
      tags: [intake]
      summary: Start an application pinned to an active template version
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "201":
          description: Created application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "409":
          description: Version is not active
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /applications/{applicationId}:
    get:
      operationId: getApplication
      tags: [intake]
      summary: Application with its intake state and pinned template
      parameters:
        - $ref: "#/components/parameters/ApplicationId"
      responses:
        "200":
          description: Application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
    patch:
      operationId: updateApplication
      tags: [intake]
      summary: Update portal-owned application fields
      description: >
        projectedClosingDate is a first-class portal-owned field (analyzer spec §8) — the
        hard-expiry clocks are judged against it. Null clears it.
      parameters:
        - $ref: "#/components/parameters/ApplicationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationUpdate"
      responses:
        "200":
          description: Updated application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
...
  /applications/{applicationId}/fields/{blockId}:
    put:
      operationId: saveFieldValues
      tags: [intake]
      summary: Save the values of a field-group block
      parameters:
        - $ref: "#/components/parameters/ApplicationId"
        - $ref: "#/components/parameters/BlockId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FieldValues"
      responses:
        "200":
          description: Updated application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
...
  /applications/{applicationId}/merge-resolutions:
    post:
      operationId: recordMergeResolution
      tags: [approved-docs]
      summary: Record (or change) the human decision on an analyzer merge recommendation
      description: >
        A merge recommendation (two run groups filed under the same block) must
        be resolved — merged or dismissed — before either group can be approved.
        Upsert keyed by run + the two page ranges; latest decision wins, so a
        dismissed recommendation can be revisited. Portal-owned, additive.
      parameters:
        - $ref: "#/components/parameters/ApplicationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MergeResolutionInput"
      responses:
        "200":
          description: Updated application (mergeResolutions upserted)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "400":
          description: Invalid body or ranges
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
        "404":
          description: Application not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
...
  /applications/{applicationId}/events:
    get:
      operationId: listApplicationEvents
      tags: [ledger]
      summary: Append-only event ledger for one application
      description: >
        Newest first. Every state change (uploads, gate decisions, verdicts,
        approvals, run ingests, template repins…) writes one row in the same
        transaction as the change. Rows are never updated or deleted.
      parameters:
        - $ref: "#/components/parameters/ApplicationId"
      responses:
        "200":
          description: Events, newest first
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListApplicationEventsResponse"
        "404":
          description: Application not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
...
components:
  parameters:
...
    ApplicationId:
      name: applicationId
      in: path
      required: true
      schema:
        type: string
    BlockId:
      name: blockId
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    ApiMessage:
      type: object
      required: [error]
      properties:
        error:
          type: string
    ApplicationEvent:
      type: object
      description: One ledger row — append-only, written with the state change it records.
      required: [seq, at, actor, action]
      properties:
        seq: { type: number }
        at: { type: string }
        actor:
          type: object
          required: [kind]
          properties:
            kind:
              type: string
              enum: [user, system]
            name: { type: string }
            ip: { type: string }
        action:
          type: string
          description: noun.verb, e.g. file.uploaded, gate.confirmed, verdict.accepted
        target:
          type: object
          required: [type]
          properties:
            type: { type: string }
            id: { type: string }
            label: { type: string }
        detail:
          type: object
          additionalProperties: true
    ListApplicationEventsResponse:
      type: array
      items:
        $ref: "#/components/schemas/ApplicationEvent"
...
    Template:
      type: object
      required: [template, version, status, program, alternatives, sections]
      properties:
        template: { type: string }
        version: { type: integer }
        status:
          type: string
          enum: [draft, active]
        program: { type: string }
        alternatives:
          type: array
          items:
            $ref: "#/components/schemas/AlternativeGroup"
        sections:
          type: array
          items:
            $ref: "#/components/schemas/Section"
...
    FieldValues:
      type: object
      description: Map of field id to entered value for one field-group block.
      required: [values]
      properties:
        values:
          type: object
          additionalProperties:
            type: string
    UploadedFile:
      type: object
      required: [filename, size, uploadedAt]
      properties:
        filename: { type: string }
        size: { type: integer }
        uploadedAt: { type: string }
        variantId:
          type: string
          description: >
            Set blocks only — which of the block's variants this file belongs to.
            Metadata only; object-store keys are unchanged.
        uploaderIp:
          type: string
          description: audit trail — client IP the upload arrived from
    ApplicationVariant:
      type: object
      description: >
        One real-world instance of a set block's requirement on THIS application
        ("Chase ····1234", "Jane Doe · 1990-04-24"). The template declares the
        descriptor shape; the application holds the actual variants. Id minted
        once, never re-minted — uploads and (later) approvals key on it.
      required: [id, descriptor, label, createdAt]
      properties:
        id: { type: string }
        descriptor:
          type: object
          description: descriptorField key -> value, keys exactly as the template block declares
          additionalProperties:
            type: string
        label: { type: string, description: "server-built display label (descriptor values joined)" }
        createdAt: { type: string }
...
    DocumentApproval:
      type: object
      description: >
        A per-DOCUMENT human approval (the filing decision) — the unit between
        page decisions and block verdicts. Tagged to a block and, for set
        blocks, a variant. approved/approved_incomplete materialize into the
        approved registry through the same seam as block accepts;
        approved_incomplete additionally requests a new version. Append-only on
        the application; a re-approval of the same pages supersedes via the
        registry, never deletes.
      required: [id, blockId, runId, pages, outcome, decidedBy, decidedAt]
      properties:
        id: { type: string }
        blockId: { type: string }
        variantId: { type: string, description: "set blocks — which variant this document files into" }
        runId: { type: string, description: "analyzer run whose grouping this approval is based on" }
        pages:
          type: array
          description: inclusive 1-based packet page range [first, last]
          items: { type: integer }
          minItems: 2
          maxItems: 2
        pageDecisions:
          type: array
          items:
            $ref: "#/components/schemas/PageDecision"
        outcome:
          type: string
          enum: [approved, approved_incomplete, rejected]
        decidedBy:
          type: string
          enum: [Originator, Underwriter, Manager]
        decidedAt: { type: string }
        approvedDocId: { type: string, description: "registry row created by materialization (approved outcomes only)" }
        pageRanges:
          type: array
          description: >
            For a document assembled from a human-accepted merge of
            NON-ADJACENT ranges: every inclusive [first,last] range that makes
            up the document, ascending, non-overlapping. `pages` remains the
            envelope [min,max] for old readers. Absent = single contiguous range.
          items:
            type: array
            items: { type: integer }
            minItems: 2
            maxItems: 2
...
    MergeResolution:
      type: object
      description: >
        The human decision on an analyzer merge recommendation between two run
        groups. Keyed on the application by "<runId>:p<f>-<l>|p<f>-<l>"
        (ranges sorted by first page). Latest decision wins — reversible.
      required: [runId, ranges, decision, decidedBy, decidedAt]
      properties:
        runId: { type: string }
        ranges:
          type: array
          description: the two inclusive page ranges the recommendation covers
          items:
            type: array
            items: { type: integer }
            minItems: 2
            maxItems: 2
          minItems: 2
          maxItems: 2
        decision:
          type: string
          enum: [merged, dismissed]
        decidedBy:
          type: string
          enum: [Originator, Underwriter, Manager]
        decidedAt: { type: string }
    MergeResolutionInput:
      type: object
      required: [runId, ranges, decision, decidedBy]
      properties:
        runId: { type: string }
        ranges:
          type: array
          items:
            type: array
            items: { type: integer }
            minItems: 2
            maxItems: 2
          minItems: 2
          maxItems: 2
        decision:
          type: string
          enum: [merged, dismissed]
        decidedBy:
          type: string
          enum: [Originator, Underwriter, Manager]
...
    ApplicationInput:
      type: object
      required: [family, version, applicantName]
      properties:
        family: { type: string }
        version: { type: integer }
        applicantName: { type: string, minLength: 1 }
        projectedClosingDate: { type: string }
    ApplicationUpdate:
      type: object
      properties:
        projectedClosingDate:
          oneOf:
            - type: string
            - type: "null"
    ApplicationSummary:
      type: object
      required: [id, family, version, templateName, applicantName, createdAt, docsFiled, docsTotal]
      properties:
        id: { type: string }
        family: { type: string }
        version: { type: integer }
        templateName: { type: string }
        applicantName: { type: string }
        createdAt: { type: string }
        docsFiled: { type: integer }
        docsTotal: { type: integer }
        projectedClosingDate: { type: string }
    Application:
      type: object
      required: [id, family, version, applicantName, createdAt, fieldValues, uploads, template]
      properties:
        id: { type: string }
        family: { type: string }
        version: { type: integer }
        applicantName: { type: string }
        createdAt: { type: string }
        fieldValues:
          type: object
          description: blockId -> field values map
          additionalProperties:
            type: object
            additionalProperties:
              type: string
        uploads:
          type: object
          description: blockId -> uploaded files
          additionalProperties:
            type: array
            items:
              $ref: "#/components/schemas/UploadedFile"
        projectedClosingDate: { type: string }
        verdicts:
          type: object
          description: blockId -> latest human verdict
          additionalProperties:
            $ref: "#/components/schemas/Verdict"
        packet:
          $ref: "#/components/schemas/PacketState"
        packetManifest:
          $ref: "#/components/schemas/PacketManifest"
        template:
          $ref: "#/components/schemas/Template"
        templateHistory:
          type: array
          description: Audit trail of template re-pins (who, when, vN→vN)
          items:
            $ref: "#/components/schemas/TemplateRepinEvent"
        materializationErrors:
          type: object
          description: >
            blockId -> last approval-materialization failure (loud, with reason).
            The verdict stands; the approved registry row is missing until a
            successful retry clears the entry. Portal-owned.
          additionalProperties:
            type: object
            required: [message, at]
            properties:
              message: { type: string }
              at: { type: string }
        documentApprovals:
          type: array
          description: >
            Per-document approvals from the filmstrip flow, append-only, newest
            first. The registry (approved-docs) is the materialized truth; this
            is the decision trail on the application. Portal-owned.
          items:
            $ref: "#/components/schemas/DocumentApproval"
        mergeResolutions:
          type: object
          description: >
            Merge-recommendation decisions, keyed "<runId>:p<f>-<l>|p<f>-<l>".
            A PENDING recommendation (no entry) gates approval of both groups.
            Portal-owned.
          additionalProperties:
            $ref: "#/components/schemas/MergeResolution"
        variants:
          type: object
          description: >
            blockId -> variants of that set block on this application. Intake-side
            data — variants and their uploads are NOT part of the application's
            satisfied requirements until a human accepts.
          additionalProperties:
            type: array
            items:
              $ref: "#/components/schemas/ApplicationVariant"
        manualPlacements:
          type: array
          description: Human filings of analyzer-unassigned page ranges (portal-owned)
          items:
            $ref: "#/components/schemas/ManualPlacement"
...
    TemplateRepinEvent:
      type: object
      required: [fromVersion, toVersion, decidedBy, decidedAt]
      properties:
        fromVersion: { type: integer }
        toVersion: { type: integer }
        decidedBy: { type: string }
        decidedAt: { type: string }
    Verdict:
      type: object
      description: >
        Human verdict on a document block. Portal-owned; the analyzer never writes these.
        documentDate/expiryDate are the confirmed dates the block's clocks run on.
      required: [verdict, datesEdited, decidedAt, decidedBy]
      properties:
        verdict:
          type: string
          enum: [accepted, new_version_requested]
        note: { type: string }
        documentDate: { type: string }
        expiryDate: { type: string }
        datesEdited: { type: boolean }
        decidedAt: { type: string }
        decidedBy:
          type: string
          enum: [Originator, Underwriter, Manager]
        runId: { type: string }
...
    ManualPlacement:
      type: object
      description: >
        A human decision about an unassigned page range: filed into a document block,
        or archived as not-relevant. Waits for the next analyzer run to confirm —
        manual placement always wins over analyzer suggestions.
      required: [pages, target, decidedBy, decidedAt]
      properties:
        pages:
          type: array
          description: inclusive 1-based page range [first, last]
          items:
            type: integer
          minItems: 2
          maxItems: 2
        target:
          type: string
          description: document blockId on the pinned template, or the literal "archive"
        note: { type: string }
        decidedBy:
          type: string
          enum: [Originator, Underwriter, Manager]
        decidedAt: { type: string }
        runId: { type: string }
...
    PacketPreflight:
      type: object
      description: >
        Deterministic pre-flight report (analyzer spec §3) — no model calls, no image
        enhancement ("gate, don't retouch"). Checks actually run in v1: file validity,
        page count, metadata snapshot, per-page blank detection, per-page contrast,
        exact-duplicate pages, embedded-image DPI. Blur/skew scoring is analyzer-tier and
        intentionally NOT claimed here. Stored on the application as audit-trail material.
      required: [verdict, flags, estimateUsd, estimateMinutes, metadata, thumbnails]
      properties:
        verdict:
          type: string
          description: One plain-language line summarising the packet's pre-flight outcome.
        flags:
          type: array
          description: Plain-language red flags ("p.6 blank", "p.4 duplicate of p.3").
          items: { type: string }
        estimateUsd:
          type: number
          description: >
            FULL-pipeline estimate (parse + judge + deep scans) — informed consent before
            spend; the judge dominates at high page counts. Staff-facing only.
        estimateMinutes: { type: integer }
        metadata:
          $ref: "#/components/schemas/PacketMetadata"
        thumbnails:
          type: array
          description: 2 worst-scoring pages + 1 best, by deterministic per-page scores.
          items:
            $ref: "#/components/schemas/PacketThumbnail"
    PacketGateDecision:
      type: object
      required: [decision, decidedAt]
      properties:
        decision:
          type: string
          enum: [auto, confirmed, bypassed]
        decidedBy:
          type: string
          description: Absent when decision=auto; otherwise the signed-in staff profile.
        decidedAt: { type: string }
...
    PacketState:
      type: object
      description: >
        Portal-owned packet state machine — the staged C2 intake flow: preflight_running →
        gated → processing → report. Persisted server-side so the gate physically blocks;
        no client-side choreography can advance it. The spec §3 auto rule (<20 clean
        pages auto-proceed) is currently suspended: every packet gates so staff can
        pick the run's model plan before spend.
      required: [filename, sizeBytes, pages, sha256, uploadedAt, state]
      properties:
        filename: { type: string }
        sizeBytes: { type: integer }
        pages: { type: integer }
        sha256: { type: string }
        uploadedAt: { type: string }
        state:
          type: string
          enum: [preflight_running, gated, processing, report]
        preflight:
          $ref: "#/components/schemas/PacketPreflight"
        gate:
          $ref: "#/components/schemas/PacketGateDecision"
        lastRunError:
          type: string
          description: >
            Plain-language reason the last analyzer kick failed (packet reverted to gated).
            Cleared by the next successful upload, gate decision, or run ingest.
        uploaderIp:
          type: string
          description: audit trail — client IP the packet upload arrived from
        files:
          type: array
          description: >
            Multi-file provenance (assembled packets only): each source file's
            name and its [first,last] span in the packet's global page numbering.
            Absent on single-PDF uploads.
          items:
            type: object
            required: [filename, pages]
            properties:
              filename: { type: string }
              pages:
                type: array
                items: { type: integer }
                minItems: 2
                maxItems: 2
    PacketManifestFile:
      type: object
      required: [id, filename, sizeBytes, pages, flags, removed]
      properties:
        id: { type: string }
        filename: { type: string }
        sizeBytes: { type: integer }
        pages: { type: integer }
        flags:
          type: array
          description: deterministic per-file quality flags (same checks as pre-flight, no model calls)
          items: { type: string }
        removed: { type: boolean }
    PacketManifest:
      type: object
      description: >
        Multi-file intake staging (Phase 5 v1): the reviewable file list built before any
        analysis. Lives on the application until assemble replaces the packet; assembling
        marks it consumed. Removed files are simply skipped at assemble — no defer queue.
      required: [files, createdAt]
      properties:
        files:
          type: array
          items:
            $ref: "#/components/schemas/PacketManifestFile"
        createdAt: { type: string }
        assembledAt:
          type: string
          description: set when assemble consumed this manifest (staging bytes are gone)
...